Switches to Zipkin v2 library
This updates to the recent `io.zipkin.reporter2` library, which sends
in v2 format by default, except scribe which is forever pinned to v1.

The net effect is this allows modern reporting libraries to be used,
such as those that exist in the `zipkin-aws` and `zipkin-gcp` repos.

This also updates the minimum Kafka version to 0.10, as it uses a more
recent Kafka library than the v0.8 version before.

Java classes take the same Finagle arguments, but are packaged in v2
conventions, as io.zipkin2.finagle. This loud change allows folks to
know things have moved.

The maven group ID is similar, but not exactly the same. It has changed
from `io.zipkin.finagle` to `io.zipkin.finagle2` as it is conventional
for zipkin group IDs to change with a version suffix on the third part
when breaking (Note: io.zipkin doesn't change as group IDs are DNS
names!)
