+++ .mvn/wrapper/MavenWrapperDownloader.java
@@ -17,9 +17,6 @@ specific language governing permissions and limitations
 under the License.
 */
 
-import java.net.*;
-import java.io.*;
-import java.nio.channels.*;
 import java.util.Properties;
 
 public class MavenWrapperDownloader {
