+++ core/pom.xml
@@ -18,9 +18,9 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.zipkin.finagle</groupId>
+    <groupId>io.zipkin.finagle2</groupId>
     <artifactId>zipkin-finagle-parent_2.12</artifactId>
-    <version>1.2.8-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>zipkin-finagle_2.12</artifactId>
@@ -32,48 +32,9 @@
 
   <build>
     <plugins>
-      <!-- Repackage internal zipkin classes -->
-      <plugin>
-        <artifactId>maven-shade-plugin</artifactId>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <minimizeJar>true</minimizeJar>
-              <createDependencyReducedPom>false</createDependencyReducedPom>
-              <relocations>
-                <relocation>
-                  <pattern>zipkin.reporter.internal</pattern>
-                  <shadedPattern>zipkin.finagle.internal</shadedPattern>
-                </relocation>
-              </relocations>
-              <artifactSet>
-                <includes>
-                  <include>io.zipkin.reporter:zipkin-reporter</include>
-                </includes>
-              </artifactSet>
-              <filters>
-                <filter>
-                  <!-- Shade references to AwaitableCallback until it is not internal -->
-                  <artifact>io.zipkin.reporter:zipkin-reporter</artifact>
-                  <includes>
-                    <include>zipkin/reporter/internal/AwaitableCallback.class</include>
-                  </includes>
-                  <excludes>
-                    <exclude>*</exclude>
-                  </excludes>
-                </filter>
-              </filters>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
       <plugin>
         <artifactId>maven-jar-plugin</artifactId>
-        <version>3.0.2</version>
+        <version>3.1.0</version>
         <executions>
           <execution>
             <goals>
