+++ http/pom.xml
@@ -18,9 +18,9 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.zipkin.finagle</groupId>
+    <groupId>io.zipkin.finagle2</groupId>
     <artifactId>zipkin-finagle-parent_2.12</artifactId>
-    <version>1.2.8-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>zipkin-finagle-http_2.12</artifactId>
@@ -46,12 +46,12 @@
       <groupId>${project.groupId}</groupId>
       <artifactId>zipkin-finagle_${scala.binary.version}</artifactId>
       <type>test-jar</type>
+      <scope>test</scope>
     </dependency>
 
     <dependency>
-      <groupId>io.zipkin.java</groupId>
+      <groupId>io.zipkin.zipkin2</groupId>
       <artifactId>zipkin-junit</artifactId>
-      <version>${zipkin.version}</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
