+++ kafka/pom.xml
@@ -18,9 +18,9 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.zipkin.finagle</groupId>
+    <groupId>io.zipkin.finagle2</groupId>
     <artifactId>zipkin-finagle-parent_2.12</artifactId>
-    <version>1.2.8-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>zipkin-finagle-kafka_2.12</artifactId>
@@ -37,9 +37,8 @@
     </dependency>
 
     <dependency>
-      <groupId>io.zipkin.reporter</groupId>
-      <artifactId>zipkin-sender-kafka08</artifactId>
-      <version>${zipkin-reporter.version}</version>
+      <groupId>io.zipkin.reporter2</groupId>
+      <artifactId>zipkin-sender-kafka11</artifactId>
     </dependency>
 
     <dependency>
