+++ pom.xml
@@ -17,9 +17,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>io.zipkin.finagle</groupId>
+  <groupId>io.zipkin.finagle2</groupId>
   <artifactId>zipkin-finagle-parent_2.12</artifactId>
-  <version>1.2.8-SNAPSHOT</version>
+  <version>2.0.0-SNAPSHOT</version>
   <packaging>pom</packaging>
 
   <modules>
@@ -46,8 +46,7 @@
     <!-- version numbers that are used more than once -->
     <scala.binary.version>2.12</scala.binary.version>
     <finagle.version>18.6.0</finagle.version>
-    <zipkin.version>2.8.4</zipkin.version>
-    <zipkin-reporter.version>1.1.2</zipkin-reporter.version>
+    <zipkin-reporter.version>2.7.6</zipkin-reporter.version>
     <license-maven-plugin.version>2.11</license-maven-plugin.version>
   </properties>
 
@@ -105,6 +104,14 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
+        <groupId>io.zipkin.reporter2</groupId>
+        <artifactId>zipkin-reporter-bom</artifactId>
+        <version>${zipkin-reporter.version}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+
+      <dependency>
         <groupId>${project.groupId}</groupId>
         <artifactId>zipkin-finagle_${scala.binary.version}</artifactId>
         <version>${project.version}</version>
@@ -133,15 +140,13 @@
     </dependency>
 
     <dependency>
-      <groupId>io.zipkin.java</groupId>
+      <groupId>io.zipkin.zipkin2</groupId>
       <artifactId>zipkin</artifactId>
-      <version>${zipkin.version}</version>
     </dependency>
 
     <dependency>
-      <groupId>io.zipkin.reporter</groupId>
+      <groupId>io.zipkin.reporter2</groupId>
       <artifactId>zipkin-reporter</artifactId>
-      <version>${zipkin-reporter.version}</version>
     </dependency>
 
     <!-- don't worry. auto-* are source retention, not runtime -->
